requirements.txt
Django==2.2.5
gunicorn==20.0.4
line-bot-sdk==1.15.0
